Improve terminal, shell, prompt, console choices
Most Git users are using a shell via terminal now and there's no reason to limit the audience of this article to windows users using out dated software.

Git CMD is a windows only option considered 2nd to Git Bash by a lot of folks, both of which have been replaced by PowerShell, and/or WSL shells on Windows.

Linux and Mac users typically use bash and zsh shells.

I don't think the audience should be limited to Windows, and if the example shell/prompt is going to be limited to one example, I think there are many better/more popular options to use as an example than Git CMD.

Shell/Prompt would probably be sufficient but the full list conveys the message that it doesn't matter how you get there as long as you can run git.